+++ tests/integration/cases/call_scalar_args_with_null/Nim_heterogeneous_strategy_record_call@v2.nim
@@ -1,0 +1,5 @@
+{.warning[UnusedImport]:off.}
+import tables
+template process(args: varargs[untyped]) = discard
+process(nil)
+process("hello")