+++ tests/integration/cases/call_transform_no_wrapper/Nim_heterogeneous_strategy_record_call@v2.nim
@@ -1,0 +1,6 @@
+{.warning[UnusedImport]:off.}
+import tables
+template process(args: varargs[untyped]) = discard
+process("hello")
+process(42)
+process(true)