+++ tests/integration/cases/dict_mixed_int_widths/Nim_heterogeneous_strategy_record@v2.nim
@@ -1,0 +1,11 @@
+{.warning[UnusedImport]:off.}
+import tables
+type Record0 = object
+    a: int
+    b: int
+    c: string
+var my_data = Record0(
+    a: 1,
+    b: 3000000000,
+    c: "x"
+)