+++ tests/integration/cases/dict_with_list_value/Nim_heterogeneous_strategy_record_list_val@v2.nim
@@ -1,0 +1,13 @@
+{.warning[UnusedImport]:off.}
+import tables
+type Record0 = object
+    name: string
+    scores: seq[int]
+var my_data = Record0(
+    name: "Alice",
+    scores: @[
+        10,
+        20,
+        30
+    ]
+)