+++ tests/integration/cases/nested_sequences/Nim_heterogeneous_strategy_record@v2.nim
@@ -1,0 +1,6 @@
+{.warning[UnusedImport]:off.}
+import tables
+var my_data = @[
+    @[@[1, 2], @[3, 4]],
+    @[@[5]]
+]