+++ tests/integration/cases/record_list_of_records/Nim_heterogeneous_strategy_record@v2.nim
@@ -1,0 +1,21 @@
+{.warning[UnusedImport]:off.}
+import tables
+type Record1 = object
+    id: int
+    label: string
+type Record0 = object
+    name: string
+    items: seq[Record1]
+var my_data = Record0(
+    name: "box",
+    items: @[
+        Record1(
+            id: 1,
+            label: "first"
+        ),
+        Record1(
+            id: 2,
+            label: "second"
+        )
+    ]
+)