+++ tests/integration/cases/record_pure_scalars/Nim_heterogeneous_strategy_record@v2.nim
@@ -1,0 +1,13 @@
+{.warning[UnusedImport]:off.}
+import tables
+type Record0 = object
+    name: string
+    age: int
+    active: bool
+    score: float
+var my_data = Record0(
+    name: "Alice",
+    age: 30,
+    active: true,
+    score: 4.5
+)